+++ zoo/src/test/scala/com/intel/analytics/zoo/serving/PostProcessingSpec.scala
@@ -17,6 +17,7 @@
 package com.intel.analytics.zoo.serving
 
 import com.intel.analytics.bigdl.tensor.{Storage, Tensor}
+import com.intel.analytics.zoo.serving.postprocessing.PostProcessing
 import org.scalatest.{FlatSpec, Matchers}
 
 class PostProcessingSpec extends FlatSpec with Matchers {
